+++ packages/google_maps_flutter/google_maps_flutter/ios/Classes/GoogleMapController.m
@@ -86,6 +86,7 @@
       }
     }];
     _mapView.delegate = weakSelf;
+    _mapView.paddingAdjustmentBehavior = kGMSMapViewPaddingAdjustmentBehaviorNever;
     _registrar = registrar;
     _markersController = [[FLTMarkersController alloc] initWithMethodChannel:_channel
                                                                      mapView:_mapView