+++ packages/camera/camera_windows/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.2.0
+
+**BREAKING CHANGES**:
+  * `CameraException.code` now has value `"CameraAccessDenied"` if camera access permission was denied.
+  * `CameraException.code` now has value `"camera_error"` if error occurs during capture.
+
+## 0.1.0+5
+
+* Fixes bugs in in error handling.
+
+## 0.1.0+4
+
+* Allows retrying camera initialization after error.
+
 ## 0.1.0+3
 
 * Updates the README to better explain how to use the unendorsed package.